backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
Django==3.2.18
django-admin==2.0.2
django-cors-headers==3.14.0
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-six==1.0.5
djangorestframework==3.14.0
idna==3.4
importlib-metadata==4.13.0
kombu==5.2.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
prompt-toolkit==3.0.38
pycparser==2.21
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
twilio==7.16.3
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
xlwt==1.3.0
zipp==3.15.0
gunicorn==20.1.0
